requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1
email-validator==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
